+++ router/core/graph_server.go
@@ -1599,9 +1599,10 @@ func (s *graphServer) buildGraphMux(
 		ApolloRouterCompatibilityFlags:                         s.apolloRouterCompatibilityFlags,
 		DisableExposingVariablesContentOnValidationError:       s.engineExecutionConfiguration.DisableExposingVariablesContentOnValidationError,
 		RelaxSubgraphOperationFieldSelectionMergingNullability: s.engineExecutionConfiguration.RelaxSubgraphOperationFieldSelectionMergingNullability,
-		EnableDefer:      s.engineExecutionConfiguration.EnableDefer,
-		ComplexityLimits: s.securityConfiguration.ComplexityLimits,
-		CostControl:      s.securityConfiguration.CostControl,
+		EnableDefer:             s.engineExecutionConfiguration.EnableDefer,
+		ComplexityLimits:        s.securityConfiguration.ComplexityLimits,
+		CostControl:             s.securityConfiguration.CostControl,
+		ValidateInlineArguments: s.engineExecutionConfiguration.ValidateInlineArguments,
 	})
 
 	if opts.ReloadPersistentState.inMemoryPlanCacheFallback.IsEnabled() {